refactor: extract session metrics schema into shared worker-utils package
Move the Zod schema and types from cloudflare-o11y into
@kilocode/worker-utils so both cloudflare-o11y and
cloudflare-session-ingest import from a single source of truth
instead of maintaining duplicate type definitions.
