+++ cloudflare-session-ingest/src/env.ts
@@ -1,1 +1,3 @@
+import type { O11YBinding } from './o11y-binding.js';
+
 export type Env = Omit<Cloudflare.Env, 'O11Y'> & { O11Y: O11YBinding };
