+++ packages/worker-utils/src/index.ts
@@ -47,3 +47,6 @@ export {
   KILO_TOKEN_VERSION,
 } from './kilo-token.js';
 export type { KiloTokenPayload, SignKiloTokenExtra } from './kilo-token.js';
+
+export { SessionMetricsParamsSchema, TerminationReasons } from './session-metrics-schema.js';
+export type { SessionMetricsParams, SessionMetricsParamsInput } from './session-metrics-schema.js';
